Move new self-healing tests to old files
