+++ tests/integration/high_availability/test_self_healing_process_killed.py
@@ -4,88 +4,101 @@
 import logging
 import time
 
-from pytest_operator.plugin import OpsTest
+import jubilant_backports
+import pytest
+from jubilant_backports import Juju
 
-from ..helpers import (
-    get_primary_unit,
-    get_process_pid,
+from constants import CONTAINER_NAME
+
+from ..helpers import generate_random_string
+from .high_availability_helpers_new import (
+    CHARM_METADATA,
+    check_mysql_units_writes_increment,
+    exec_k8s_container_command,
+    get_mysql_primary_unit,
+    get_unit_process_id,
+    insert_mysql_test_data,
+    remove_mysql_test_data,
+    update_interval,
+    verify_mysql_test_data,
+    wait_for_apps_status,
 )
-from .high_availability_helpers import (
-    clean_up_database_and_table,
-    ensure_all_units_continuous_writes_incrementing,
-    ensure_n_online_mysql_members,
-    get_application_name,
-    insert_data_into_mysql_and_validate_replication,
-    send_signal_to_pod_container_process,
-)
 
-logger = logging.getLogger(__name__)
+MYSQL_APP_NAME = "mysql-k8s"
+MYSQL_PROCESS_NAME = "mysqld"
+MYSQL_TEST_APP_NAME = "mysql-test-app"
 
-MYSQL_CONTAINER_NAME = "mysql"
-MYSQLD_PROCESS_NAME = "mysqld"
-TIMEOUT = 40 * 60
+MINUTE_SECS = 60
+
+logging.getLogger("jubilant.wait").setLevel(logging.WARNING)
 
 
-async def test_kill_db_process(
-    ops_test: OpsTest, highly_available_cluster, continuous_writes, credentials
-) -> None:
+@pytest.mark.abort_on_fail
+def test_deploy_highly_available_cluster(juju: Juju, charm: str) -> None:
+    """Simple test to ensure that the MySQL and application charms get deployed."""
+    logging.info("Deploying MySQL cluster")
+    juju.deploy(
+        charm=charm,
+        app=MYSQL_APP_NAME,
+        base="ubuntu@22.04",
+        config={"profile": "testing"},
+        resources={"mysql-image": CHARM_METADATA["resources"]["mysql-image"]["upstream-source"]},
+        num_units=3,
+    )
+    juju.deploy(
+        charm=MYSQL_TEST_APP_NAME,
+        app=MYSQL_TEST_APP_NAME,
+        base="ubuntu@22.04",
+        channel="latest/edge",
+        config={"sleep_interval": 300},
+        num_units=1,
+    )
+
+    juju.integrate(
+        f"{MYSQL_APP_NAME}:database",
+        f"{MYSQL_TEST_APP_NAME}:database",
+    )
+
+    logging.info("Wait for applications to become active")
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, MYSQL_APP_NAME, MYSQL_TEST_APP_NAME
+        ),
+        error=jubilant_backports.any_blocked,
+        timeout=20 * MINUTE_SECS,
+    )
+
+
+@pytest.mark.abort_on_fail
+def test_kill_db_process(juju: Juju, continuous_writes) -> None:
     """Test to send a SIGKILL to the primary db process and ensure that the cluster self heals."""
-    mysql_application_name = get_application_name(ops_test, "mysql")
+    logging.info("Ensuring all units have continuous writes incrementing")
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
 
-    logger.info("Waiting until 3 mysql instances are online")
-    # ensure all units in the cluster are online
-    assert await ensure_n_online_mysql_members(ops_test, 3), (
-        "The deployed mysql application is not fully online"
+    mysql_primary_unit = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    mysql_primary_unit_pid = get_unit_process_id(juju, mysql_primary_unit, MYSQL_PROCESS_NAME)
+
+    logging.info(f"Killing process id {mysql_primary_unit_pid}")
+    exec_k8s_container_command(
+        juju=juju,
+        unit_name=mysql_primary_unit,
+        container_name=CONTAINER_NAME,
+        command="pkill -f mysqld --signal SIGKILL",
     )
 
-    logger.info("Ensuring all units have continuous writes incrementing")
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
-
-    mysql_unit = ops_test.model.applications[mysql_application_name].units[0]
-    primary = await get_primary_unit(ops_test, mysql_unit, mysql_application_name)
-
-    mysql_pid = await get_process_pid(
-        ops_test, primary.name, MYSQL_CONTAINER_NAME, MYSQLD_PROCESS_NAME
-    )
-
-    logger.info(f"Sending SIGKILL to unit {primary.name}")
-    await send_signal_to_pod_container_process(
-        ops_test.model.info.name,
-        primary.name,
-        MYSQL_CONTAINER_NAME,
-        MYSQLD_PROCESS_NAME,
-        "SIGKILL",
-    )
-
-    # Wait for the SIGKILL above to take effect before continuing with test checks
     time.sleep(10)
 
-    logger.info("Waiting until 3 mysql instances are online")
-    assert await ensure_n_online_mysql_members(ops_test, 3), (
-        "The mysql application is not fully online after sending SIGKILL to primary"
-    )
+    new_mysql_primary_unit_pid = get_unit_process_id(juju, mysql_primary_unit, MYSQL_PROCESS_NAME)
+    assert new_mysql_primary_unit_pid != mysql_primary_unit_pid
 
-    # ensure that the mysqld process got restarted and has a new process id
-    new_mysql_pid = await get_process_pid(
-        ops_test, primary.name, MYSQL_CONTAINER_NAME, MYSQLD_PROCESS_NAME
-    )
-    assert mysql_pid != new_mysql_pid, (
-        "The mysql process id is the same after sending it a SIGKILL"
-    )
+    # Ensure continuous writes still incrementing for all units
+    with update_interval(juju, "10s"):
+        check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
 
-    new_primary = await get_primary_unit(ops_test, mysql_unit, mysql_application_name)
-    assert primary.name != new_primary.name, (
-        "The mysql primary has not been reelected after sending a SIGKILL"
-    )
+    # Ensure that we are able to insert data into the primary
+    table_name = "data"
+    table_value = generate_random_string(255)
 
-    logger.info("Ensuring all units have continuous writes incrementing")
-    # ensure continuous writes still incrementing for all units
-    async with ops_test.fast_forward():
-        await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
-
-    # ensure that we are able to insert data into the primary and have it replicated to all units
-    database_name, table_name = "test-kill-db-process", "data"
-    await insert_data_into_mysql_and_validate_replication(
-        ops_test, database_name, table_name, credentials
-    )
-    await clean_up_database_and_table(ops_test, database_name, table_name, credentials)
+    insert_mysql_test_data(juju, MYSQL_APP_NAME, table_name, table_value)
+    verify_mysql_test_data(juju, MYSQL_APP_NAME, table_name, table_value)
+    remove_mysql_test_data(juju, MYSQL_APP_NAME, table_name)
