+++ tests/integration/high_availability/test_self_healing_stop_primary.py
@@ -3,89 +3,94 @@
 
 import logging
 
-from pytest_operator.plugin import OpsTest
-from tenacity import Retrying, stop_after_delay, wait_fixed
+import jubilant_backports
+import pytest
+from jubilant_backports import Juju
 
-from ..helpers import (
-    get_primary_unit,
-    get_process_pid,
+from constants import CONTAINER_NAME
+
+from .high_availability_helpers_new import (
+    CHARM_METADATA,
+    check_mysql_units_writes_increment,
+    exec_k8s_container_command,
+    get_mysql_primary_unit,
+    get_unit_process_id,
+    wait_for_apps_status,
 )
-from .high_availability_helpers import (
-    ensure_all_units_continuous_writes_incrementing,
-    ensure_n_online_mysql_members,
-    get_application_name,
-    send_signal_to_pod_container_process,
-)
 
-logger = logging.getLogger(__name__)
+MYSQL_APP_NAME = "mysql-k8s"
+MYSQL_PROCESS_NAME = "mysqld"
+MYSQL_TEST_APP_NAME = "mysql-test-app"
 
-MYSQL_CONTAINER_NAME = "mysql"
-MYSQLD_PROCESS_NAME = "mysqld"
-TIMEOUT = 40 * 60
+MINUTE_SECS = 60
+
+logging.getLogger("jubilant.wait").setLevel(logging.WARNING)
 
 
-async def test_graceful_crash_of_primary(
-    ops_test: OpsTest, highly_available_cluster, continuous_writes, credentials
-) -> None:
+@pytest.mark.abort_on_fail
+def test_deploy_highly_available_cluster(juju: Juju, charm: str) -> None:
+    """Simple test to ensure that the MySQL and application charms get deployed."""
+    logging.info("Deploying MySQL cluster")
+    juju.deploy(
+        charm=charm,
+        app=MYSQL_APP_NAME,
+        base="ubuntu@22.04",
+        config={"profile": "testing"},
+        resources={"mysql-image": CHARM_METADATA["resources"]["mysql-image"]["upstream-source"]},
+        num_units=3,
+    )
+    juju.deploy(
+        charm=MYSQL_TEST_APP_NAME,
+        app=MYSQL_TEST_APP_NAME,
+        base="ubuntu@22.04",
+        channel="latest/edge",
+        config={"sleep_interval": 300},
+        num_units=1,
+    )
+
+    juju.integrate(
+        f"{MYSQL_APP_NAME}:database",
+        f"{MYSQL_TEST_APP_NAME}:database",
+    )
+
+    logging.info("Wait for applications to become active")
+    juju.wait(
+        ready=wait_for_apps_status(
+            jubilant_backports.all_active, MYSQL_APP_NAME, MYSQL_TEST_APP_NAME
+        ),
+        error=jubilant_backports.any_blocked,
+        timeout=20 * MINUTE_SECS,
+    )
+
+
+@pytest.mark.abort_on_fail
+async def test_graceful_crash_of_primary(juju: Juju, continuous_writes) -> None:
     """Test to send SIGTERM to primary instance and then verify recovery."""
-    mysql_application_name = get_application_name(ops_test, "mysql")
+    # Ensure continuous writes still incrementing for all units
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
 
-    assert mysql_application_name, "mysql application name is not set"
+    mysql_primary_unit = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    mysql_primary_unit_pid = get_unit_process_id(juju, mysql_primary_unit, MYSQL_PROCESS_NAME)
 
-    logger.info("Ensuring that there are 3 online mysql members")
-    assert await ensure_n_online_mysql_members(ops_test, 3), (
-        "The deployed mysql application does not have three online nodes"
+    logging.info(f"Terminating process id {mysql_primary_unit_pid}")
+    exec_k8s_container_command(
+        juju=juju,
+        unit_name=mysql_primary_unit,
+        container_name=CONTAINER_NAME,
+        command=f"pkill -f {MYSQL_PROCESS_NAME} --signal SIGTERM",
     )
 
-    logger.info("Ensuring that all units have incrementing continuous writes")
-    await ensure_all_units_continuous_writes_incrementing(ops_test, credentials=credentials)
+    new_mysql_primary_unit_pid = get_unit_process_id(juju, mysql_primary_unit, MYSQL_PROCESS_NAME)
+    assert new_mysql_primary_unit_pid == mysql_primary_unit_pid
 
-    mysql_unit = ops_test.model.applications[mysql_application_name].units[0]
-    primary = await get_primary_unit(ops_test, mysql_unit, mysql_application_name)
-
-    mysql_pid = await get_process_pid(
-        ops_test, primary.name, MYSQL_CONTAINER_NAME, MYSQLD_PROCESS_NAME
+    logging.info("Waiting until there are 3 online mysql instances again")
+    juju.wait(
+        ready=wait_for_apps_status(jubilant_backports.all_active, MYSQL_APP_NAME),
+        timeout=20 * MINUTE_SECS,
     )
 
-    logger.info(f"Sending SIGTERM to unit {primary.name}")
-    await send_signal_to_pod_container_process(
-        ops_test.model.info.name,
-        primary.name,
-        MYSQL_CONTAINER_NAME,
-        MYSQLD_PROCESS_NAME,
-        "SIGTERM",
-    )
+    new_mysql_primary_unit = get_mysql_primary_unit(juju, MYSQL_APP_NAME)
+    assert new_mysql_primary_unit != mysql_primary_unit
 
-    new_mysql_pid = await get_process_pid(
-        ops_test, primary.name, MYSQL_CONTAINER_NAME, MYSQLD_PROCESS_NAME
-    )
-    assert new_mysql_pid == mysql_pid, (
-        "mysql process id is not the same as it was before process was stopped"
-    )
-
-    remaining_online_units = [
-        unit
-        for unit in ops_test.model.applications[mysql_application_name].units
-        if unit.name != primary.name
-    ]
-
-    logger.info("Waiting until there are 3 online mysql instances again")
-    # retrying as it may take time for the cluster to recognize that the primary process is stopped
-    for attempt in Retrying(stop=stop_after_delay(2 * 60), wait=wait_fixed(10)):
-        with attempt:
-            assert await ensure_n_online_mysql_members(ops_test, 3), (
-                "The deployed mysql application does not have three online nodes"
-            )
-
-            new_primary = await get_primary_unit(
-                ops_test, remaining_online_units[0], mysql_application_name
-            )
-            assert primary.name != new_primary.name, "new mysql primary was not elected"
-
-    logger.info("Ensuring all instances have incrementing continuous writes")
-    async with ops_test.fast_forward():
-        for attempt in Retrying(stop=stop_after_delay(60), wait=wait_fixed(10)):
-            with attempt:
-                await ensure_all_units_continuous_writes_incrementing(
-                    ops_test, credentials=credentials
-                )
+    # Ensure continuous writes still incrementing for all units
+    check_mysql_units_writes_increment(juju, MYSQL_APP_NAME)
