--- tests/spread/test_crash_during_setup.py/task.yaml
@@ -1,7 +1,0 @@
-summary: test_crash_during_setup.py
-environment:
-  TEST_MODULE: high_availability/test_crash_during_setup.py
-execute: |
-  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
-artifacts:
-  - allure-results
