--- tests/spread/test_node_drain.py/task.yaml
@@ -1,7 +1,0 @@
-summary: test_node_drain.py
-environment:
-  TEST_MODULE: high_availability/test_node_drain.py
-execute: |
-  tox run -e integration -- "tests/integration/$TEST_MODULE" --model testing --alluredir="$SPREAD_TASK/allure-results"
-artifacts:
-  - allure-results
